refactor: move day entity to the core domain
