+++ lib/core/domain/entities/day.dart
