+++ lib/core/domain/entities/day.freezed.dart
