+++ lib/core/domain/repositories/day_repository.dart
@@ -1,0 +1,5 @@
+import 'package:hidroly/core/domain/entities/day.dart';
+
+abstract class DayRepository {
+  void save(Day day);
+}
