--- lib/features/setup/domain/repositories/day_repository.dart
@@ -1,5 +1,0 @@
-import 'package:hidroly/features/setup/domain/entities/day.dart';
-
-abstract class DayRepository {
-  void save(Day day);
-}
